+++ apps/docs/src/app/core/api-files.ts
@@ -14,6 +14,12 @@ export const API_FILES = {
         'ActionBarHeaderDirective',
         'ActionBarMobileDirective'
     ],
+    actionSheet: [
+        'ActionSheetComponent',
+        'ActionSheetControlComponent',
+        'ActionSheetItemComponent',
+        'ActionSheetBodyComponent'
+    ],
     alert: ['AlertComponent', 'AlertConfig', 'AlertService', 'AlertRef'],
     avatar: ['AvatarComponent'],
     bar: ['BarComponent', 'BarElementDirective', 'BarLeftDirective', 'BarMiddleDirective', 'BarRightDirective'],
@@ -52,6 +58,7 @@ export const API_FILES = {
         'DateFormatParserDefault',
         'FdDate'
     ],
+    carousel: ['CarouselComponent', 'CarouselItemComponent', 'CarouselService'],
     combobox: ['ComboboxComponent'],
     checkbox: ['CheckboxComponent'],
     datePicker: ['DatePickerComponent', 'DateFormatParser', 'DateFormatParserDefault', 'FdDate'],
@@ -82,7 +89,7 @@ export const API_FILES = {
     infiniteScroll: ['InfiniteScrollDirective'],
     inlineHelp: ['InlineHelpComponent'],
     inputGroup: ['InputGroupComponent'],
-    layoutGrid: ['LayoutGridComponent', 'LayoutGridSpanDirective'],
+    layoutGrid: ['LayoutGridComponent', 'LayoutGridColDirective', 'LayoutGridRowDirective'],
     layoutPanel: [
         'LayoutPanelComponent',
         'LayoutPanelActionsComponent',
@@ -283,6 +290,7 @@ export const API_FILES = {
         'TimeColumnComponent'
     ],
     timePicker: ['TimePickerComponent', 'TimeFormatParser', 'TimeFormatParserDefault', 'TimeObject'],
+    title: ['TitleComponent'],
     switch: ['SwitchComponent'],
     stepInput: ['StepInputComponent'],
     token: ['TokenComponent', 'TokenizerComponent'],