+++ apps/docs/src/app/core/component-docs/action-sheet/examples/action-sheet-mobile/action-sheet-mobile-example.component.ts
@@ -1,0 +1,11 @@
+import {
+    ChangeDetectionStrategy,
+    Component
+} from '@angular/core';
+
+@Component({
+    selector: 'fd-action-sheet-mobile-example',
+    templateUrl: './action-sheet-mobile-example.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ActionSheetMobileExampleComponent {}