+++ apps/docs/src/app/core/component-docs/carousel/examples/carousel-error-message.component.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-carousel-error-message',
+    templateUrl: './carousel-error-message.component.html',
+    styles: ['fd-carousel:focus {outline: 1px dotted}', 'img {pointer-events: none;}'],
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class CarouselErrorMessageComponent {}