+++ apps/docs/src/app/core/component-docs/layout-grid/examples/index.ts
@@ -1,0 +1,24 @@
+import { ExampleLayoutGridBlockComponent } from './example-layout-grid-block.component';
+import {
+    LayoutGridBasicExampleComponent,
+    LayoutGridNestingExampleComponent,
+    LayoutGridNoGapExampleComponent,
+    LayoutGridOffsetExampleComponent,
+    LayoutGridResponsiveExampleComponent,
+    LayoutGridResponsiveOffsetExample,
+    LayoutGridRowExampleComponent
+} from './layout-grid-examples.component';
+
+export * from './layout-grid-examples.component';
+export * from './example-layout-grid-block.component';
+
+export const examples = [
+    ExampleLayoutGridBlockComponent,
+    LayoutGridRowExampleComponent,
+    LayoutGridBasicExampleComponent,
+    LayoutGridNoGapExampleComponent,
+    LayoutGridOffsetExampleComponent,
+    LayoutGridNestingExampleComponent,
+    LayoutGridResponsiveOffsetExample,
+    LayoutGridResponsiveExampleComponent
+];