+++ apps/docs/src/app/core/component-docs/popover/examples/popover-simple/popover-example.component.ts
@@ -1,8 +1,4 @@
 import { Component, ViewEncapsulation } from '@angular/core';
-import { RtlService } from '@fundamental-ngx/core';
-import { Observable } from 'rxjs';
-import { Placement } from 'popper.js';
-import { map } from 'rxjs/operators';
 
 @Component({
     selector: 'fd-popover-example',
@@ -11,14 +7,6 @@ import { map } from 'rxjs/operators';
     encapsulation: ViewEncapsulation.None
 })
 export class PopoverExampleComponent {
-    leftPlacement$: Observable<Placement>;
-    rightPlacement$: Observable<Placement>;
-
-    constructor(private _rtlService: RtlService) {
-        this.leftPlacement$ = this._rtlService.rtl.pipe(map((isRtl) => (isRtl ? 'right' : 'left')));
-        this.rightPlacement$ = this._rtlService.rtl.pipe(map((isRtl) => (isRtl ? 'left' : 'right')));
-    }
-
     list1 = [
         { text: 'Option 1', url: '#' },
         { text: 'Option 2', url: '#' },