+++ apps/docs/src/app/core/component-docs/title/title-docs.module.ts
@@ -1,0 +1,39 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ApiComponent } from '../../../documentation/core-helpers/api/api.component';
+import { SharedDocumentationPageModule } from '../../../documentation/shared-documentation-page.module';
+import { API_FILES } from '../../api-files';
+import { TitleHeaderComponent } from './title-header/title-header.component';
+import { TitleDocsComponent } from './title-docs.component';
+import { TitleModule } from '@fundamental-ngx/core';
+import {
+    TitleElisionExampleComponent,
+    TitleSemanticExampleComponent,
+    TitleVisualExampleComponent,
+    TitleWrappingExampleComponent
+} from './examples/title-examples.component';
+
+const routes: Routes = [
+    {
+        path: '',
+        component: TitleHeaderComponent,
+        children: [
+            { path: '', component: TitleDocsComponent },
+            { path: 'api', component: ApiComponent, data: { content: API_FILES.title } }
+        ]
+    }
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes), SharedDocumentationPageModule, TitleModule],
+    exports: [RouterModule],
+    declarations: [
+        TitleDocsComponent,
+        TitleHeaderComponent,
+        TitleSemanticExampleComponent,
+        TitleElisionExampleComponent,
+        TitleVisualExampleComponent,
+        TitleWrappingExampleComponent
+    ]
+})
+export class TitleDocsModule {}