+++ apps/docs/src/app/core/core-documentation.routes.ts
@@ -17,6 +17,11 @@ export const ROUTES: Routes = [
                     import('./component-docs/action-bar/action-bar-docs.module').then((m) => m.ActionBarDocsModule)
             },
             {
+                path: 'action-sheet',
+                loadChildren: () =>
+                    import('./component-docs/action-sheet/action-sheet-docs.module').then((m) => m.ActionSheetDocsModule)
+            },
+            {
                 path: 'alert',
                 loadChildren: () => import('./component-docs/alert/alert-docs.module').then((m) => m.AlertDocsModule)
             },
@@ -59,6 +64,11 @@ export const ROUTES: Routes = [
                 path: 'calendar',
                 loadChildren: () =>
                     import('./component-docs/calendar/calendar-docs.module').then((m) => m.CalendarDocsModule)
+            },
+            {
+                path: 'carousel',
+                loadChildren: () =>
+                    import('./component-docs/carousel/carousel-docs.module').then((m) => m.CarouselDocsModule)
             },
             {
                 path: 'checkbox',
@@ -182,6 +192,10 @@ export const ROUTES: Routes = [
             {
                 path: 'list',
                 loadChildren: () => import('./component-docs/list/list-docs.module').then((m) => m.ListDocsModule)
+            },
+            {
+                path: 'list-byline',
+                loadChildren: () => import('./component-docs/list-byline/list-byline-docs.module').then((m) => m.ListDocsModule)
             },
             {
                 path: 'localizationEditor',
@@ -348,6 +362,10 @@ export const ROUTES: Routes = [
                 path: 'timePicker',
                 loadChildren: () =>
                     import('./component-docs/time-picker/time-picker-docs.module').then((m) => m.TimePickerDocsModule)
+            },
+            {
+                path: 'title',
+                loadChildren: () => import('./component-docs/title/title-docs.module').then((m) => m.TitleDocsModule)
             },
             {
                 path: 'token',