+++ libs/core/src/lib/action-sheet/action-sheet-mobile/action-sheet-mobile.component.ts
@@ -1,0 +1,26 @@
+import {
+    Component,
+    Input,
+    TemplateRef
+} from '@angular/core';
+
+@Component({
+    selector: 'fd-action-sheet-mobile',
+    templateUrl: './action-sheet-mobile.component.html'
+})
+export class ActionSheetMobileComponent {
+
+    /** Whenever links should be visible **/
+    @Input()
+    open = false;
+
+    /** @hidden */
+    childContent: {
+        actionSheetBodyTemplate: TemplateRef<any>,
+    } = null;
+
+    /** @hidden */
+    close(): void {
+        this.open = false;
+    }
+}