+++ libs/core/src/lib/list/directives/byline/list-thumbnail.directive.ts
@@ -1,0 +1,10 @@
+import { Directive, HostBinding } from '@angular/core';
+
+@Directive({
+    selector: '[fdListThumbnail], [fd-list-thumbnail]'
+})
+export class ListThumbnailDirective {
+    /** @hidden */
+    @HostBinding('class.fd-list__thumbnail')
+    fdListThumbnailClass = true;
+}