+++ libs/core/src/lib/notification/notification-service/notification.service.spec.ts
@@ -1,6 +1,8 @@
 import { TestBed } from '@angular/core/testing';
 import { NgModule, TemplateRef, Component, ViewChild } from '@angular/core';
-import { NotificationModule, NotificationRef, NotificationService } from '@fundamental-ngx/core';
+import { NotificationService } from '../notification-service/notification.service';
+import { NotificationRef } from '../notification-utils/notification-ref';
+import { NotificationModule } from '../notification.module';
 
 @Component({
     template: `